refactor(memory): split the compartment fan-out out of pipeline.py
